requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
sqlparse==0.4.4
tzdata==2023.3
pytz==2023.3
whitenoise==6.6.0
gunicorn==21.2.0